refactor: rewrite fourcc load and convert
Squashed commit of the following:

commit 65026ecc22957e5becf79d1719f592c22e8d1641
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 14:06:35 2023 +0800

    refactor: use PixelFormat class for better readability

commit fcd032d934c978005bc46d3d9356c6d9b8325764
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 13:30:55 2023 +0800

    cleanup: delete some useless lien of code

commit 8dc352a0643f6bd58ee31d026dac7c6763e83e03
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 13:24:35 2023 +0800

    refactor: reuse file reading for various fourcc images

commit 27bd6d94809d60d676c7513d9bc85a9b1c936d9b
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 12:12:20 2023 +0800

    clean: use precise var name `mapped_ext`

commit 28fc033083ac9fdc4ef698b9c4ca29bcb0b18691
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 11:54:00 2023 +0800

    Use all zero values for displayed image when load non-image file

commit 5c924d98c79bf80ff40d155cf2b4a2366204fd50
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 11:47:24 2023 +0800

    clean: convert std::cout loggging to LOG(level)

commit 1d0119899035a7bb3815fe168931909048cf57b7
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 04:19:46 2023 +0800

    Use xkcd-script font

commit cd9ba828a9b65813242a645fff53b852c91b6bfe
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 03:24:20 2023 +0800

    Update cmake plugins, fix warning reported by overlook

+++ src/kantu/string.cpp
@@ -3,7 +3,7 @@
 std::string kantu::to_lower(const std::string& str)
 {
     std::string lower_str = str;
-    for (int i = 0; i < str.size(); i++)
+    for (size_t i = 0; i < str.size(); i++)
     {
         if (isalpha(str[i]))
         {
